engine/requirements.txt
fastapi
uvicorn
playwright
beautifulsoup4
google-cloud-firestore
google-cloud-storage
vertexai
pydantic
python-dotenv
aiohttp
